Add some code to deal with RSA private key imports
This code is needed to calculate the coefficients needed for faster
calculation that nettle requires when importing a private key.

The code is not "safe", it can spill intermediate values all over the
heap, and it is only a stop gap until cryptography safe code can be
found to replace this.

Signed-off-by: Simo Sorce <simo@redhat.com>