Initial scaffold for a token agnostic test app
This is supposed to load an arbitrary token and run a battery of tests
printing out PASS/FAIL output.

The idea is that we can load arbitrary tokens and see how much they
conform to some baseline and test that the cryptographic functions they
support actually perform as expected.

Additionally it should be able to execute test vectors to check
conformance of cryptographic routines.

Signed-off-by: Simo Sorce <simo@redhat.com>