Make a more versatile OsslParam object
There are several benefits to using a full object here:
- encapsulates several unsafe aspects in a controlled manner.
- avoids the need to carry around auxiliary vectors to store data in an
unsafe manner (the previous code had no borrow guarantees, the vec could
be freed earlier and the params struct would end up with dangling
pointers).
- provides a more streamlined API to quickly create OSSL_PARAM arrays.

Additionally this allows to remove a bunch of macro_rules, which makes
for better debuggability.

Signed-off-by: Simo Sorce <simo@redhat.com>