kasn1: Strip leading zeroes when creating bignums
When creating the ASN.1 structure from PKCS#11 signature, we can have
leading zeroes that make the rust-asn1 fail to accept such encoding as
it is not minimal.
This was not an issue for P-256 curves, but are much more freqent with
P-521 signatures, where the PKCS#11 buffer is fixed and almost
always will start with leading zero.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>