Ensure we flush data when it changes
In various cases changeing data in the cache would not cause a flush to
the json file. Writes are rare so even for the very inefficient json
file storage it is ok to just flush (and therefore rewrite the entire
file) all the changes as soon as any modification operation happens.

Signed-off-by: Simo Sorce <simo@redhat.com>