Add sqlite storage
Add an sqlite storage backend and change the existing code to become a
memory and a json backend (with the latter using the former as temporary
storage).

Change token to address caching storage changes, which require a
mutable object to update the cache on reads as well. Introduce
the conept of CheckedHandle to force correct use of the API.
It needed to be split into a "prefetch" call which takes a mutable
reference to the storage, and an actual call to get the object that does
not.
This is because the borrow checkers holds a mutable reference to the
objects whenever a reference (even immutable) is returned as a result.
This then prevents making further calls to the token object.
Once theborrow checker is improved to stop inferring incorrect things
about the need to keep mutable references around, this can be changed
back to be a slightly saner access pattern.

Change test infrastructure to use the files in testdata/ to generate
separate test files that are always removed after the test completes.
Switch most tests to use the sql storage.

Signed-off-by: Simo Sorce <simo@redhat.com>