Add PBKDF2 support
This is another "good one" where PKCS#11 spec deviates from what you
would expect, and uses C_GenerateKey to implement Key derivation instead
of C_DeriveKey.

This means we now need to pass mechanisms and object factories to the
generate function because the template can specify any key type.
So we need to get the generate_key function from the PBKDF mechanism,
but we need to be able to find the create() mechanism for any supported
key type ... and we need to source the HMAC mechanism ...

Signed-off-by: Simo Sorce <simo@redhat.com>