Fix attribute types to unbreak loading token info
The load_token_info() references the token info attributes as strings,
but they are defined as bytes, which makes the checks in
get_attr_as_string() fail when reading from the storage (there is
similar check also on the sqlite storage backed).

We decided to create vendor specific attributes to make sure we do not
clash with any other definition in pkcs11 and did explicit
adding/removal of the trailing spaces.

This was not caught by the testsuite as it does not re-initialize the
token, but keeps using the token after token initialization, which never
triggers the reload of the token info from the database.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>