tests: Reproducer for issues reading tokeninfo from the storage
The test can not be executed in parallel with any other tests in the
thread as it finalizes the pkcs11 module, which breaks any other running
sessions in the same process. Therefore the annotation of all the tests
as explicitly parallel and keeping these two new serial only.

This is tested separately with both json and sql backend as they had
different issues in them.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>