Make OpenSSL context wrappers safer
Instead of exposing raw pointers which is quite unsafe,
internalize as many unsafe operations as possible on these
raw pointers so that the actual interface of these wrapped
type is mostly safe.

Declare interfaces that still deal with raw pointers as unsafe so
that proper care is done when handling them.

Signed-off-by: Simo Sorce <simo@redhat.com>