Add support for loading a default token
This token is always localted in slot 0, so if a slot already exists
there an error will be returned.
By default searches in order for a file in:
KRYOPTIC_CONF
XDG_HOME_DIR/kryoptic/token.sql
HOME/.local/share/kryoptic/token.sql
/var/krypoptic/public/token.sql

The last is just in case the system wants to provide some sort of
default read only token, prhaps with public certificates or so.

Signed-off-by: Simo Sorce <simo@redhat.com>