Allow to pick a free slot when not specified
A user can specify a specific slot number to use when initializiing a
token. However if a filename (or env var) is passed that does not
specify a token and slot number 0 is not available, allow to use the
first available free slot as long as the same filename was not already
used to initialze a token.

Signed-off-by: Simo Sorce <simo@redhat.com>