Add support for CKA_ALWAYS_AUTHENTICATE behavior
When a private key has this attribute set to true it commands that any
operation (decryption/signature/etc..) done with this key requires an
extra authentication to authorize the individual operation.

This is done by setting flags on the session that will control proper
authentication is checked if necessary, and support for the special
CKU_CONTEXT_SPECIFIC user type into the fn_login() function.

Signed-off-by: Simo Sorce <simo@redhat.com>