Refactor derive operation return value
Instead of concocting a complex set of calls to return keys one by one,
just live with the fact lib.rs needs to do some mechanism specific
handling for those few derive operation that need to fill parameters
with key handles.

This simplifies the multiple aspects of the derive Operation, although
the amount of code is about the same.

Signed-off-by: Simo Sorce <simo@redhat.com>