Use safe casting of integer types
The 'as' construct does blind casting and no checks on whether the
resulting data is sane. It is especially important on the FFI boundaries
to be extra careful about what inputs and outputs we ingest/emit.

Use the TryFrom trait to cast numerical types and error out
appropriately whenever something silly happens to be found.

Signed-off-by: Simo Sorce <simo@redhat.com>