Fix a casting error
Proper bound checked casting paid off by discovering this issue where
we were blindingly storing a CK_ULONG::MAX in the DB that we can't
properly decode back as the max integer we can represent is a i64.

This fix ensures the storage format is properly addressing > u32::MAX
values which should never be allowed in CK_ULONG attribues.

Signed-off-by: Simo Sorce <simo@redhat.com>