Use AES Key Wrap for wrapping keys
Using GCM to wrap keys is not really the right thing to do and due
to various limitations in FIPS mode it can fail to work as the IV should
always be random in that case. FIPS also favors the use of KW functions
for key wrapping, so just use that.

Signed-off-by: Simo Sorce <simo@redhat.com>