Add configuration file support
To aid cases where an environment variable cannot be easily set nor
init arguments can be easily passed we need a place where we can store
the tokens configuration, including the ability to specify multiple
slots.

The ability to use init args and ignore any configuration is retained but
the init arguments format has been enhanced to be a set of comma
separated key/value pairs so that additional info can be passed there
in the future.

Tests:
Add test to verify multiple tokens in config

This will try to load a configuration that specifies multiple tokens.

Adjust all tests that depend on the env var to run serially so they do
not interfeere with each other and actually force load the config.

Signed-off-by: Simo Sorce <simo@redhat.com>