Refactor Storage Layer
This splits out the storage spefcific knowledge from the token
to the actual storage backend.

The storage backend is restructured so that multiple storage types can
use a common information organizational schema for non standardized
objects like user pins and token info and non standardized
authenticationand encryption layers for the actual data.

This is in preparation for adding code to read databases written by
other software tokens that follow different storage philosophies.

While woring on this refactoring a few minor issues with how pin/login
were handled was found and have been fixed.

Additionally the code now uses a common Key encryption Key that both the
Ueer and SO Pins can unlock allowing recovery of the database via SO
pin if the user pin is lost. This means the SO pin needs to be set to a
non-default value on initialization to maintain the confidentaility of
the token.
It can be set to a random value if recovery is never desired.

Signed-off-by: Simo Sorce <simo@redhat.com>