Correct slot assignment
Using 0 as the default slot number was masking cases when the
configuration may not be correctly imported.

By using an "invalid" value by default we ensure that a slot number has
been properly assigned, and better check that there are no duplicate
(same backing database) tokens loaded in different slots.

Signed-off-by: Simo Sorce <simo@redhat.com>