Add support for initializing nssdbs
NSSDBs have sepcific lists of attributes they deal with, and although
technically you can grow tables, we need to be able to handle attributes
that are NSS specific but may not care about in Kryoptic.

Handle initialization from scrath as well as writing data back to the
db.

At the same time we will need to filter out attributes Kryoptic may use
that are not supported by NSS, spin off more files to keep individual
files to reasonable sizes.

Signed-off-by: Simo Sorce <simo@redhat.com>