Ensure CKK_EC_EDWARDS public keys are cheked
The initial implementation of Edwards keys used a definition partially
consisent with the language in the 3.0 spec that has since changed in
the 3.1 spec.

Namely the 3.0 spec indicated that the public CKA_EC_POINT for
CKK_EC_EDWARDS curves needed to be encoded as a DER BIT STRING, however
the implementation was taking a DER OCTET STRING instead.

The 3.1 spec clarifes that the CKA_EC_POINT for Edwards curves is just a
buffer of bytes, with no ASN.1 encoding, just like the private
CKA_VALUE.

This patch adds all the necessary checks on import and aligns the
implementation to the 3.1 Spec.

Signed-off-by: Simo Sorce <simo@redhat.com>