Handle compatibility for CKA_EC_POINT
The PKCS#11 3.0 Spec was released with a bug in the description of the
public point encoding for Edwards and Montgomery curves that foooled
many implementations to assume these curves would return a DER encoded
octet string as the value for CKA_EC_POINT.

However an errata was issued and the 3.1 spec also confirms that the
CKA_EC_POINT for these curves should be just a byte array (like the
private value is).

It is easy to accept either form and conver it for storage, the real
tricky part is returning the point to applications. Applications that
only handle one form and can't deal with different encodings may get
an encoding they do not like and fail.

Signed-off-by: Simo Sorce <simo@redhat.com>