ci: Skip memorydb tests
The test drivers are now written for persistent storage to properly
excercise the whole process from library initialization, reading
data from persistent storage, writing them and more.

The memorydb is specific as it is not persistent. In the tests,
we are able to initialize it successfully, but the content is thrown
away and then a new empty token is created again, which does not
have the expected content.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>