nssdb: Handle skipped attributes when querying
The SKIP_ATTRIBUTES are not stored in the NSS DB. When constructing the
queries, these need to be skipped otherwise they will reference
non-existing table columns and fail on the DB level.

But the application is asking for them so we need to mock the default
values back when returning from fetch().

This demonstrated when destroying a token objects, which checks if the
object has the CKA_DESTROYABLE attribute.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>