Set a ceiling to the number of keys in cache
This is a very simple way to avoid unbounded memory pressure, however it
has no smarts to keep in cache the most used keys. It simply kicks out
the last key whch is the key that happens to have the "higher" salt.
It can cause performance degradation if a token has very many keys and
keeps trying to access a roration of those with a "higher" salt.

Signed-off-by: Simo Sorce <simo@redhat.com>