Replace Zeroize with OPENSSL_cleanse
The Zeroize crate performance is crazy slow.
The search for a performance issue with the nssdb storage via
valgrind's callgrind revelead that 50% of the time in the HMAC code was
spent on zeroization because of the way the Zeroize crate is built.

Replace it everywhere with a simpler zeromem function that underneath
simply calls OPENSSL_cleanse() on the mutable slice passed in.

With this change alone the new test that performs 100 pbkdf2 calls went
from ~9s to ~4.5s on my laptop.

Signed-off-by: Simo Sorce <simo@redhat.com>