aes: Check tag length for FIPS indicators
The check_mac_fips_indicators() won't do anything for the CMACs,
as the cmac implementation in FIPS provider does not implement this
parameter.

The AES key with SIGN/VERIFY flags was also not marked correctly as FIPS
valid so any CMAC operation on that given key caused the FIPS indicator
failure.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>