Fix AES Key-Wrap checks
Key wrap checks operate on the wrapped key, which is of arbitrary type,
so we switch CKK_AES for KRY_UNSPEC without enforcing any size.

Because any key can be wrapped we also need to allow all flags in
genflags as any of them can be set on unwrapping.

Finally fix ECC key size checks, we store he key size in bytes so we
need to check against the len in bytes not in bits; convert bits to
bytes when sourcing keylen for ECC keys.

Signed-off-by: Simo Sorce <simo@redhat.com>