aes: Unify encrypt/decrypt initialization
The functions encrypt_initialize() and decrypt_initialize() have most of
the duplicated code except for the OpenSSL function names. This unfifies
the use on functions that work in both direction, keeping the code
simpler and needing to fix bugs only in one place.

There were slight differences already in both of the implementations so
I tried to take the better variant with more descriptive comments and
checking.

Fixes: #145

Signed-off-by: Jakub Jelen <jjelen@redhat.com>