Silly way to generate cscope database
As we have to use OUT_DIR, there is no simple way to index the generated
files as they are stored in a location that changes frequently due to the
hash in the target directory.

Use a blunt search for the file name of each of the generated files and
take the one with the latest modification date.

Signed-off-by: Simo Sorce <simo@redhat.com>