Change how operations are handled
PKCS#11 allows multiple operations of different types to be executed in
parallel on a token. Drop the single operation enum method, and instead
allow storage for each type of operation on the session.

This also improves compile times checks, and drops runtime checks to
check that the correct operation type is being accessed.

Fixes #168

Signed-off-by: Simo Sorce <simo@redhat.com>