Switch to 2024 edition
The migration tool automatically changed some of thcode causing some
churn. I manually changed some of the things, like removing expr_2021
from macros as it was not necessary and changing the renaming of the now
reserved word gen from r#gen to generator.

A bunch of let Some() {} else () expressions were changed to matches, I
let that be, sometime smatch turns out to be idiomatic, sometimes it is
about the same readability.

Signed-off-by: Simo Sorce <simo@redhat.com>