Change how [en/de]cryption and length checks work
Change the [en/de]cryption_len() functions semantics.
a) allow calling these functions regardless of internal status, as now
they are not tied to it.

b) When called with the fin value set to true, the functions now return
the total needed output buffer without assuming the caller is at the
final step. The function does include internal state in the computation
but can be called even immediately after initialization.

This means that fn_encrypt()/fn_decrypt() functions can easily query
for the total required length, as well as _final() functions with the
same code.

c) When called without the 'fin' boolean turned on, assume that the
request is being made in the context of calling just the next _update()
function. This means calculations only enforce the minimum buffer output
needed for the input data being injested and does not need to account
for additional data that some modes my require on final like AES_CBC_PAD,
AES_GCM, and others that append padding or tags.

d) For block ciphers allow providing partial inputs, as long as, before
calling final(), the correct size has been fed into the operation we
are now able to take partial input buffers even for modes that require
the total input to be block sized.

e) An internal buffer is always used (for block modes) to store partial
blocks, and OpenSSL is only ever asked to encrypt block-sized quantities.
Modes that allow non-block sized inputs but require block sized outputs
complete the last short write in the _final() operation.

f) The above does require the caller to always provide enough output
buffer size to take entire output blocks taking in account possible
partial blocks previously fed to the function. The new
[en/de]cryption_len() semantics always correctly report how much output
is required in these cases.

Signed-off-by: Simo Sorce <simo@redhat.com>