Enable the FIPS Indicator callback
This patch is more convoluted than usual because OpenSSL calls the
provider-provided get_libctx() callback to obtain the libctx to pass to
the indicator callback.

It does this in the self-tests too, which are called from
OSSL_provider_init_int(), that is before initialization of FIPS_PROVIDER
is complete.

But we store the libctx on the FIPS_PROVIDER structure which is being
lazily initialized at startup causing a deadlock.

To prevent the deadlock, set a canary pointer as the core handle, that
we can check in fips_get_libctx().

If the canary pointer is not set as the core handle yet, avoid trying
to access the real libctx from the yet uninitlized and locked
FIPS_PROVIDER (via get_libctx()) and instead immediately return NULL.

Signed-off-by: Simo Sorce <simo@redhat.com>