Fix check for object sensitivity as per spec
The spec says:
"If the CKA_SENSITIVE attribute is CK_TRUE, or if the CKA_EXTRACTABLE
attribute is CK_FALSE, then certain attributes of the private key cannot
be revealed in plaintext"

However the code was performing a logical "and" instead of an "or".
Correct the code and the tests to properly check/set the CKA_SENSITIVE value.

Signed-off-by: Simo Sorce <simo@redhat.com>