Fix dependcy install issue in CI
We can't test anymore on i686 as the i686 opensc tools are not available
anymore after the split out of the opensc-libs package (which is still
available but useless to us).

So drop the i686 test for migration. We still build the i686 version
and run internal unit tests in the build target as that doesn't require
external tools.

Signed-off-by: Simo Sorce <simo@redhat.com>