Fix ML-DSA signature code with newer OpenSSL
The initial PR was created and tested only against 3.5.0 and had 2 bugs
when used against the master tree which finally implements multi-part
signatures.

The code to set signatures at initialization with detection was calling
EVP_SIGNATURE_settable_ctx_params() to check which parameters were
settable, but unfortunately this function does not pass a context
(because it doesn't have one) to the provider resulting in the default
settables only being returned. Using EVP_PKEY_CTX_settable_params()
instead provides the correct context pointer and we can correctly
detect that setting signatures is supported.

The code did not set the signature in the context when the signature
was not passed at initialization (C_VerifyInit API case).
In this case, when we detect multi-part is supported, we need to set
the signature before calling EVP_PKEY_verify_message_final().

Signed-off-by: Simo Sorce <simo@redhat.com>