Add better support for array of CK_ULONG attrs
The only attribute we currently support that is of this kind is the
CKA_ALLOWED_MECHANISMS attribute.
This code is necessary to ensure the data we store can be read by
architectures where CK_ULONG is not a 64bit value or that have
different endianess in their representation.

For backwards compatibility with the current storage format on
x86_64/arm64 we store these attributes as arrays of 64bit values in
little endian order.

For the json storage, which does not have a stable format, we just
change the format to use an array of integers.

Signed-off-by: Simo Sorce <simo@redhat.com>