Replace __va_list_tag with just c_void
Bindgen seem to generate different names for va arg lists based on which
platforms it is runnignon, specifically it geenrates a different name on
aarch64, so these declaration fail to compile as the type is unknown.

Given we never use these arguments and they are always a pointer, just
declare these functions as taking a void pointer, which makes them
compile fine on all platfroms.

See: https://github.com/rust-lang/rust-bindgen/issues/2631

Signed-off-by: Simo Sorce <simo@redhat.com>