Ensure secret_len is big enough
In a CI test in an unrelated PR we got a failure where, the
(secret_len - keylen) calculation resulted in a negative integer error.
Rust caught this as an invalid operation for an unsigned integer and
paniced.

This probably happened because openssl returns the maximum shared secret
length when we query for allocating the buffer size, but the actual length
of the computed shared secret may be slightly shorter based on what the
actual secret and public key values end up being.

Ensure that we always have a secret_len result from a call to
EVP_PKEY_derive that is usable (no smaller than the smaller key
size for AES). And compute/validate keylen after we have the final
computed secret_len value.

Finally, ensure we return exactly keylen bytes, regardless of the size
of the "secret" buffer.

Signed-off-by: Simo Sorce <simo@redhat.com>