Always log and save traces on CI failures
Instead of treating log as a sterile standalone target.
Always use it and save the trace on CI failures

Currently trace.log files end up full of leading 0s for
some reason, this is something to investigate, it may be
that the log stuff is not safe to be called from multiple C
threads at the same time.

Signed-off-by: Simo Sorce <simo@redhat.com>