Allow variance in FFDH derived key maximum size
When FFDH derives a key it may sometimes return a result that is
slightly (1 byte) shorter than the key size.

When we don't have a requested key len in the derivation template do
not hard fail if the returned result is not as long as the keysize.

Generally callers are recommended to actually pass in a valid key
length anyway, but some of our tests don't do that, and I think we
should keep them like to to test theis case anyway.

Signed-off-by: Simo Sorce <simo@redhat.com>