Reduce max key sizes to avoid overflow
On i686 the value 0xffffffff is the maximum value that a CK_ULONG can
carry, which causes an overflow when we use ($val + 7 )/ 8 to convert
bits to bytes.

Given that it is never reasonable to have keys that are gigabits long
anyway let's clam the max down to 16M bits (2MiB) for now.

Signed-off-by: Simo Sorce <simo@redhat.com>