simplekdf: Fix attribute derivation
Previously, the logic was flawed and the attributes were not properly
set. This was not noticed on the session objects as they were not
written to the database.

This reworks how the derivation works, leaves most of the default
attributes to the `default_object_derive()` function and adjusts
just the cases where we combine two keys.

It also turned out we never set the common object attributes on import
using C_CreateObject(), which is done here too.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>