Add profiles feature to RPM packaging
This enables the new `profiles` cargo feature for the RPM build process.

The `kryoptic.spec` file is refactored to use a global variable for the list
of enabled features, improving maintainability. The `quick-xml` crate, a new
dependency required by the profiles feature, is added to the GitHub Actions
build environment.

Signed-off-by: Simo Sorce <simo@redhat.com>