Add Public Certificates Token profile tests
This commit introduces support for testing the PKCS #11 "Public Certificates
Token" profile.

A new `--gencerts` option is added to the conformance tool. When used during
token initialization, it populates the token with a set of predefined
certificate objects and a corresponding CKO_DATA object. The test certificate
data is sourced from NSS.

The test executor for C_GetAttributeValue is enhanced to support both querying
for attribute length and fetching the full value. It now also validates the
returned attribute values against the expected results defined in the test
profile XML.

Finally, the test case for this profile is enabled in the test script.

Signed-off-by: Simo Sorce <simo@redhat.com>