Clear errors from signature probe
The ossl bindings check for support of message_update in the specific
signature provider by attempting an update operation. If this probe
failed because the function was unsupported, it would leave an error
on the OpenSSL error stack.

Because OpenSSL's error stack is tied to a thread local variable rather
than the libctx this can propagate to the application kryoptic is loaded
into.

This change wraps the underlying EVP calls with `ERR_set_mark` and
`ERR_pop_to_mark`. This ensures that any errors generated by the probe are
cleared, preventing them from propagating and causing spurious failures in
subsequent code.

Signed-off-by: Simo Sorce <simo@redhat.com>